refactor: move Nav Groups into a folder
